+++ backend/src/connection/connection.gateway.ts
@@ -14,12 +14,14 @@ import { Repository } from 'typeorm';
 import { corsOption } from '../common/option/cors.option';
 import { AppConfigService } from '../config/app/configuration.service';
 import { Friendship } from '../entity/friendship.entity';
-import { InvisibleChannelRepository } from '../repository/invisible-channel.repository';
-import { Channel } from '../repository/model/channel';
-import { Status } from '../repository/model/user-status';
-import { SocketIdRepository } from '../repository/socket-id.repository';
-import { UserStatusRepository } from '../repository/user-status.repository';
-import { VisibleChannelRepository } from '../repository/visible-channel.repository';
+
+import {
+  InvisibleChannelRepository,
+  SocketIdRepository,
+  UserStatusRepository,
+  VisibleChannelRepository,
+} from '../repository';
+import { Channel, Status } from '../repository/model';
 
 @WebSocketGateway({ cors: corsOption })
 export class ConnectionGateway implements OnGatewayConnection, OnGatewayDisconnect {
