+++ backend/src/repository/index.ts
@@ -1,0 +1,7 @@
+export * from './channel.repository';
+export * from './game.repository';
+export * from './invisible-channel.repository';
+export * from './invitation.repository';
+export * from './socket-id.repository';
+export * from './user-status.repository';
+export * from './visible-channel.repository';
