+++ backend/src/repository/model/index.ts
@@ -1,0 +1,5 @@
+export * from './channel';
+export * from './game';
+export * from './invitation';
+export * from './socket-id';
+export * from './user-status';
